Break pipx.util import cycle
